Add monolith version of hydrologic model
